🛡️ Sentinel: Standardize secure ID generation and enhance telemetry logging
This commit introduces a standardized, cryptographically secure ID generation
utility and hardens telemetry logging by ensuring all events pass through the
PII-redacting logger.

Changes:
- Created `src/lib/id-utils.ts` with a zero-dependency `generateSecureId` to
  avoid Cloudflare build issues in foundational modules.
- Updated `logger.ts`, `session-analytics.ts`, and `ab-test.ts` to use secure IDs.
- Re-exported `generateSecureId` from `src/lib/utils.ts` for backward compatibility.
- Replaced direct `console.log` in `session-analytics.ts` with `logger.debug` to
  ensure PII redaction is applied to analytics events.
- Verified fix with `pnpm build:cloudflare` and `npm run security:check`.
